requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
google-genai>=0.10.0
openai>=1.0.0
anthropic>=0.30.0
Pillow>=10.0.0
pillow-heif>=0.18.0
pdf2image>=1.16.0
json-repair>=0.30.0
